Refactor: 상대주소 변환 및 인덴트 정리
