+++ server/api/src/domain/auth/dto/login-request.dto.ts
@@ -1,4 +1,5 @@
 import { IsEmail, IsNotEmpty } from 'class-validator';
+
 export class LoginRequestDto {
   @IsEmail()
   @IsNotEmpty()
