+++ server/api/src/domain/history/history.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { HistoryController } from './controller/history.controller';
 import { HistoryService } from './service/history.service';
-import { UserRepository } from 'src/domain/user/repository/user.repository';
+import { UserRepository } from '../user/repository/user.repository';
 import { HistoryRepository } from './repository/history.repository';
 import { VisitRepository } from './repository/visit.repository';
