+++ server/api/src/domain/history/repository/visit.repository.ts
@@ -1,7 +1,7 @@
 import { EntityRepository, Repository } from 'typeorm';
 import { LocalDate } from 'js-joda';
 import { Visit } from '../visit.entity';
-import { VisitBuilder } from '../../../builder/index';
+import { VisitBuilder } from '../../../builder';
 
 @EntityRepository(Visit)
 export class VisitRepository extends Repository<Visit> {
