+++ server/api/src/domain/room/dto/room-response.dto.ts
@@ -1,6 +1,6 @@
 import { Room, RoomType } from '../room.entity';
 import { UserResponseDto } from '../../user/dto/user-response.dto';
-import { UserResponseDtoBuilder } from '../../../builder/user/user-response.dto.builder';
+import { UserResponseDtoBuilder } from '../../../builder';
 import { User } from '../../user/user.entity';
 
 export class RoomResponseDto {
