+++ server/api/src/scheduler/scheduler.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { Cron } from '@nestjs/schedule';
 import { InjectRepository } from '@nestjs/typeorm';
-import { UserRepository } from 'src/domain/user/repository/user.repository';
-import { VisitRepository } from 'src/domain/history/repository/visit.repository';
+import { UserRepository } from '../domain/user/repository/user.repository';
+import { VisitRepository } from '../domain/history/repository/visit.repository';
 
 @Injectable()
 export class SchedulerService {
